ADD NEW MODULE

+++ web_hendler/cleaner/cleaner.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"log"
 	"os"
+	"os/exec"
 	"path/filepath"
 	"runtime"
 	"strings"
@@ -14,7 +15,7 @@ import (
 /*
 Сканирует указаный каталог на указанную глубину, удаляет из него файлы коорые старше указанного времени
 */
-func ScanOldFile(dir string, day int, depth int) {
+func ScanOldFile(dir string, day int, depth int, platform string) {
 	log.Printf("[WEB_HENDLER][CLEANER] Провожу сканирование и удаляю файлы которые старше %d дней.\n", day)
 	list, err := GetListFile(dir, depth)
 	if err != nil {
@@ -22,7 +23,7 @@ func ScanOldFile(dir string, day int, depth int) {
 	}
 
 	for _, value := range list {
-		DeleteOldFile(value, day)
+		DeleteOldFile(value, day, platform)
 	}
 
 }
@@ -31,7 +32,7 @@ func ScanOldFile(dir string, day int, depth int) {
 Определение возраста файла, если файл старше указанного возраста то файл удалеятьеся
 Нужно указать путь и возвраст.
 */
-func DeleteOldFile(path string, day int) {
+func DeleteOldFile(path string, day int, platform string) {
 	fileInfo, err := os.Stat(path)
 	if err != nil {
 		if os.IsNotExist(err) {
@@ -46,14 +47,44 @@ func DeleteOldFile(path string, day int) {
 	age := time.Since(modeTime)
 
 	if age > (time.Duration(day) * 24 * time.Hour) {
-		fmt.Printf("[WEB_HENDLER][CLEANER] Удаляю папку или файл %s, старше %d дней \n", path, day)
+		//fmt.Printf("[WEB_HENDLER][CLEANER] Удаляю папку или файл %s, старше %d дней \n", path, day)
 		if !fileInfo.IsDir() {
-			os.Remove(path)
+			fmt.Printf("[WEB_HENDLER][CLEANER] Удаляю папку или файл %s, старше %d дней \n", path, day)
+			if strings.HasSuffix(fileInfo.Name(), "apk") || strings.HasSuffix(fileInfo.Name(), "obb") {
+				if status := cloud(fileInfo.Name(), path, platform); status == 0 {
+					fmt.Printf("[WEB_HENDLER][CLEANER] Удаляю папку или файл %s, старше %d дней \n", path, day)
+					os.Remove(path)
+				}
+			}
 		}
 	} else {
 		fmt.Printf("[WEB_HENDLER][CLEANER] Папка или файл %s, младше %d дней \n", path, day)
 
 	}
+}
+
+func cloud(name string, path string, platform string) int {
+	path_to_cloud_mod, _ := os.LookupEnv("PATH_CLOUD_MOD")
+	if path_to_cloud_mod != "" {
+		log.Println("PATH_CLOUD_MOD:", path_to_cloud_mod)
+	} else {
+		panic("Ошибка!!! Не установлен путь к исполняемому файлу модуля отправки файлов в облако")
+	}
+
+	cmdArg := []string{
+		"-name", name,
+		"-path", path,
+		"-platform", platform,
+	}
+
+	log.Println("Получены аргументы для отправки в облако: ", cmdArg)
+	cmd := exec.Command(path_to_cloud_mod, cmdArg...)
+	if err := cmd.Run(); err != nil {
+		log.Println("Ошибка при запуске модуля загрузки данных в облако", err)
+	}
+
+	return cmd.ProcessState.ExitCode()
+
 }
 
 /*
